+++ src/kernel_lion_decode.h
@@ -93,12 +93,11 @@ typedef struct {
     bool efficiencyChecked;
     bool readSignature;
 
-    uint_fast64_t endDataOverhead;
+    uint_fast8_t endDataOverhead;
 
     density_lion_form_data formData;
 
     uint_fast32_t lastHash;
-    //uint_fast8_t lastUnigram;
     uint32_t lastChunk;
 
     density_lion_dictionary dictionary;
@@ -107,7 +106,7 @@ typedef struct {
 
 static const uint8_t density_lion_decode_bitmasks[DENSITY_LION_DECODE_NUMBER_OF_BITMASK_VALUES] = DENSITY_LION_DECODE_BITMASK_VALUES;
 
-DENSITY_KERNEL_DECODE_STATE density_lion_decode_init(density_lion_decode_state *, const density_main_header_parameters parameters, const uint_fast32_t);
+DENSITY_KERNEL_DECODE_STATE density_lion_decode_init(density_lion_decode_state *, const density_main_header_parameters parameters, const uint_fast8_t);
 DENSITY_KERNEL_DECODE_STATE density_lion_decode_continue(density_memory_teleport *, density_memory_location *, density_lion_decode_state *);
 DENSITY_KERNEL_DECODE_STATE density_lion_decode_finish(density_memory_teleport *, density_memory_location *, density_lion_decode_state *);
